add logger

+++ routes/path.ts
@@ -1,21 +1,27 @@
 import pathController from "../controllers/pathController.ts";
-import db from "../utils/db.ts";
-import Server from "../models/Server.ts";
+import serversController from "../controllers/serversController.ts";
 import { Router } from "../utils/deps.ts";
 
 const router = Router();
 const pathUrl = "/:alias/:path(*)";
+const serversDb = new serversController();
 
 router.get(pathUrl, async (req, res) => {
   const { alias, path } = req.params;
-  const collection = await db.getCollection<Server>("servers");
-  const url = collection.findOne({ alias }).url as string;
-  const clientFtp = new pathController(url, 21);
-  await clientFtp.connectToServer();
+  const port = req.query.port || 21;
+  const url = (await serversDb.getOne(alias)).url;
+
+  if (!url) {
+    throw res.setStatus(404).json("server not exist.");
+  }
+  const clientFtp = new pathController(url, port);
   try {
+    console.log("ohshit", path);
     const data = await clientFtp.downloadFile(path);
+    console.log("ohshit", data);
     res.setStatus(200).send(data);
   } catch (err) {
+    console.error(err);
     if (err.code === 550) {
       res.setStatus(404).json("file or directory not exist.");
     } else {
@@ -26,13 +32,16 @@ router.get(pathUrl, async (req, res) => {
 
 router.post(pathUrl, async (req, res) => {
   const { alias, path } = req.params;
+  const port = req.query.port || 21;
   console.log("PUSH NEW FILE");
   const file = req.body.file;
   console.log(file);
-  const collection = await db.getCollection<Server>("servers");
-  const url = collection.findOne({ alias }).url as string;
-  const clientFtp = new pathController(url, 21);
-  await clientFtp.connectToServer();
+  const url = (await serversDb.getOne(alias)).url;
+  if (!url) {
+    throw res.setStatus(404).json("server not exist.");
+  }
+
+  const clientFtp = new pathController(url, port);
   const data = await clientFtp.uploadFile(path, file);
   console.log(data);
   res.setStatus(200).json({
@@ -43,12 +52,15 @@ router.post(pathUrl, async (req, res) => {
 
 router.put(pathUrl, async (req, res) => {
   const { alias, path } = req.params;
+  const port = req.query.port || 21;
   const { newName } = req.body;
-  const collection = await db.getCollection<Server>("servers");
-  const url = collection.findOne({ alias }).url as string;
-  const clientFtp = new pathController(url, 21);
-  await clientFtp.connectToServer();
-  const data = await clientFtp.renameFile(path, newName);
+  const url = (await serversDb.getOne(alias)).url;
+  if (!url) {
+    throw res.setStatus(404).json("server not exist.");
+  }
+  const clientFtp = new pathController(url, port);
+  //try catch
+  await clientFtp.renameFile(path, newName);
   res.setStatus(200).json({
     success: "true",
     data: "camarche",
@@ -57,10 +69,12 @@ router.put(pathUrl, async (req, res) => {
 
 router.delete(pathUrl, async (req, res) => {
   const { alias, path } = req.params;
-  const collection = await db.getCollection<Server>("servers");
-  const url = collection.findOne({ alias }).url as string;
-  const clientFtp = new pathController(url, 21);
-  await clientFtp.connectToServer();
+  const port = req.query.port || 21;
+  const url = (await serversDb.getOne(alias)).url;
+  if (!url) {
+    throw res.setStatus(404).json("server not exist.");
+  }
+  const clientFtp = new pathController(url, port);
   const data = await clientFtp.deleteFile(path);
   console.log(data);
   res.setStatus(200).json({
